Moved ProfileTransformer to ProfileController
