+++ app/Http/Controllers/User/ProfileController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\User;
 use Auth;
 use Illuminate\Http\JsonResponse;
 use App\Http\Controllers\Controller;
+use App\Transformers\User\ProfileTransformer;
 use App\Services\Contracts\UserProfileService;
 use App\Http\Requests\UpdateUserProfileRequest;
 
@@ -32,7 +33,7 @@ class ProfileController extends Controller
     {
         $user = $this->userProfileService->getGeneral(Auth::user()->id);
 
-        return response()->json($user);
+        return fractal($user, new ProfileTransformer())->respond();
     }
 
     /**
