+++ app/Services/Contracts/UserProfileService.php
@@ -2,6 +2,7 @@
 
 namespace App\Services\Contracts;
 
+use App\User;
 use App\Services\Requests\{
     UpdateUserAvatarRequest,
     UpdateUserProfileRequest
@@ -14,9 +15,9 @@ interface UserProfileService
      *
      * @param int $userId
      *
-     * @return array
+     * @return \App\User
      */
-    public function getGeneral(int $userId): array;
+    public function getGeneral(int $userId): User;
 
     /**
      * Update the user profile general data.
